refactor: move sidebar to common components folder
